Fix make zip on mixed python2 / python3 Linux systems, maybe others

+++ modules/rbf_read.py
@@ -1,3 +1,4 @@
+#!/usr/bin/env python2
 import struct, sys
 
 # from http://stackoverflow.com/questions/35988/c-like-structures-in-python
@@ -7,9 +8,9 @@ class Bunch:
 
 # RBF loading code from rbfEditor, rebUtils.py
 # http://freecode.com/projects/rbfeditor
-_FNT_HDR_MAGIC 	= '\xE0\x0E\xF0\x0D\x03\x00\x00\x00'
-_FNT_HDR_SIZE	= 0x74
-_FNT_MAX_NAME 	= 64
+_FNT_HDR_MAGIC  = '\xE0\x0E\xF0\x0D\x03\x00\x00\x00'
+_FNT_HDR_SIZE   = 0x74
+_FNT_MAX_NAME   = 64
 
 def rbf_load(file):
     '''
@@ -21,31 +22,31 @@ def rbf_load(file):
     file.seek(0)
     magic = file.read(len(_FNT_HDR_MAGIC))
     if magic == _FNT_HDR_MAGIC :
-        fmt = "="+str(_FNT_MAX_NAME)+'s'
-        self.name		= struct.unpack(fmt,file.read(_FNT_MAX_NAME))[0]
+        fmt = "=" + str(_FNT_MAX_NAME) + 's'
+        self.name       = struct.unpack(fmt,file.read(_FNT_MAX_NAME))[0]
         fmt = '=l'
         fmt_size = 4
-        self._charSize 	= struct.unpack(fmt,file.read(fmt_size))[0]
-        self.points 	= struct.unpack(fmt,file.read(fmt_size))[0]
-        self.height		= struct.unpack(fmt,file.read(fmt_size))[0]
-        self.maxWidth 	= struct.unpack(fmt,file.read(fmt_size))[0]
-        self.charFirst 	= struct.unpack(fmt,file.read(fmt_size))[0]
-        self._charLast	= struct.unpack(fmt,file.read(fmt_size))[0]
-        self._unknown4 	= struct.unpack(fmt,file.read(fmt_size))[0]
-        self._wmapAddr 	= struct.unpack(fmt,file.read(fmt_size))[0]
-        self._cmapAddr 	= struct.unpack(fmt,file.read(fmt_size))[0]
-        self.descent	= struct.unpack(fmt,file.read(fmt_size))[0]
-        self.intline 	= struct.unpack(fmt,file.read(fmt_size))[0]
-        self.wTable		= []
-        self.cTable		= []
+        self._charSize  = struct.unpack(fmt,file.read(fmt_size))[0]
+        self.points     = struct.unpack(fmt,file.read(fmt_size))[0]
+        self.height     = struct.unpack(fmt,file.read(fmt_size))[0]
+        self.maxWidth   = struct.unpack(fmt,file.read(fmt_size))[0]
+        self.charFirst  = struct.unpack(fmt,file.read(fmt_size))[0]
+        self._charLast  = struct.unpack(fmt,file.read(fmt_size))[0]
+        self._unknown4  = struct.unpack(fmt,file.read(fmt_size))[0]
+        self._wmapAddr  = struct.unpack(fmt,file.read(fmt_size))[0]
+        self._cmapAddr  = struct.unpack(fmt,file.read(fmt_size))[0]
+        self.descent    = struct.unpack(fmt,file.read(fmt_size))[0]
+        self.intline    = struct.unpack(fmt,file.read(fmt_size))[0]
+        self.wTable     = []
+        self.cTable     = []
 
-        self.width = 8 * self._charSize / self.height
+        self.width = int(8 * self._charSize / self.height)
         self.charCount = self._charLast - self.charFirst + 1
 
         charlist = xrange(0, self.charCount)
         file.seek(self._wmapAddr)
         for char in charlist:
-            self.wTable.append(struct.unpack('=B',file.read(1))[0])
+            self.wTable.append(struct.unpack('=B', file.read(1))[0])
         file.seek(self._cmapAddr)
         for char in charlist:
             self.cTable.append(file.read(self._charSize))
@@ -56,10 +57,10 @@ font = None
 
 def rbf_init_font(file):
     global font
-    font = rbf_load(open(file,"rb"))
+    font = rbf_load(open(file, "rb"))
 
 def extent_func(string):
     w = 0
     for c in string:
         w += font.wTable[ord(c) - font.charFirst]
-    return (w,0)
+    return (w, 0)
